Odyss.AI.Backend.OCR/requirements.txt
flask
quart
pydantic
pydantic_core
transformers
Pillow
PyPDF2
pdfplumber
datetime
python-dotenv
python-docx
reportlab
torch
tiktoken
python-Levenshtein
nltk
pytesseract
paddleocr
paddlepaddle
quart-cors
bson
pip2tex
pdf2image